// INTEL_FEATURE_ISA_AMX2 Replace uint64_t with unsigned long long in Intel_amx2intrin.h // end INTEL_FEATURE_ISA_AMX2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/194380
Tested-by: Topper, Craig <craig.topper@intel.com>
Old/split clang commit ID: a21569e0d30b0dea680c6cf99fbcc344506673b9

+++ clang/lib/Headers/Intel_amx2intrin.h
@@ -64,21 +64,21 @@
                    "r"(index))
 #define _tile_storehd(tile, base, index, scale)                                \
   __asm__ volatile("tstorehd %%tmm" #tile ", (%0,%1," #scale ")" ::"r"(base),  \
-                   "r"((uint64_t)(index)))
+                   "r"((unsigned long long)(index)))
 #define _tile_storehdt1(tile, base, index, scale)                              \
   __asm__ volatile("tstorehdt1 %%tmm" #tile ", (%0,%1," #scale                 \
                    ")" ::"r"(base),                                            \
-                   "r"((uint64_t)(index)))
+                   "r"((unsigned long long)(index)))
 #define _tile_storentd(tile, base, index, scale)                               \
   __asm__ volatile("tstorentd %%tmm" #tile ", (%0,%1," #scale ")" ::"r"(base), \
-                   "r"((uint64_t)(index)))
+                   "r"((unsigned long long)(index)))
 #define _tile_storeqd(tile, base, index, scale)                                \
   __asm__ volatile("tstoreqd %%tmm" #tile ", (%0,%1," #scale ")" ::"r"(base),  \
-                   "r"((uint64_t)(index)))
+                   "r"((unsigned long long)(index)))
 #define _tile_storeqdt1(tile, base, index, scale)                              \
   __asm__ volatile("tstoreqdt1 %%tmm" #tile ", (%0,%1," #scale                 \
                    ")" ::"r"(base),                                            \
-                   "r"((uint64_t)(index)))
+                   "r"((unsigned long long)(index)))
 #define _tile_storerowd(tile, mem)                                             \
   __asm__ volatile("tstorerowd %%tmm" #tile ", %0" ::"m"(mem))
 // Format
